requirements.txt
hikari==2.0.0.dev103
hikari-lightbulb==1.5.0
pytz==2021.1
requests==2.25.1
